# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ThinkPixel Agent Runtime API
  version: 0.1.0-draft
  description: Vendor-neutral durable Session and bounded Execution API.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
...
servers:
  - url: https://thinkpixelar.example/v1
security:
  - bearerAuth: []
paths:
  /sessions:
    post:
      operationId: createSession
      summary: Create a durable Session
      description: Creates a PROVISIONING Session with an immutable operator-approved runtime binding. The current local implementation accepts only an empty source and reserves Workspace/source operation identities; materialization is separate. Authentication and current tenant authorization are required on creation and replay.
      tags: [Sessions]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateSession'}}}}
      responses: {'201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/Session'}}}}, '202': {$ref: '#/components/responses/Accepted'}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '409': {$ref: '#/components/responses/Conflict'}, '413': {$ref: '#/components/responses/PayloadTooLarge'}, '415': {$ref: '#/components/responses/UnsupportedMediaType'}, '422': {$ref: '#/components/responses/UnprocessableContent'}, '429': {$ref: '#/components/responses/RateLimited'}, '503': {$ref: '#/components/responses/Unavailable'}, '504': {$ref: '#/components/responses/Timeout'}, default: {$ref: '#/components/responses/Problem'}}
    get:
      operationId: listSessions
      summary: List visible Sessions
      tags: [Sessions]
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses: {'200': {description: Page, content: {application/json: {schema: {$ref: '#/components/schemas/SessionPage'}}}}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '429': {$ref: '#/components/responses/RateLimited'}, '503': {$ref: '#/components/responses/Unavailable'}, default: {$ref: '#/components/responses/Problem'}}
  /sessions/{session_id}:
    parameters: [{$ref: '#/components/parameters/SessionId'}]
    get:
      operationId: getSession
      summary: Get a Session
      tags: [Sessions]
      responses: {'200': {description: Session, content: {application/json: {schema: {$ref: '#/components/schemas/Session'}}}}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '429': {$ref: '#/components/responses/RateLimited'}, '503': {$ref: '#/components/responses/Unavailable'}, default: {$ref: '#/components/responses/Problem'}}
    delete:
      operationId: closeSession
      summary: Close a Session
      tags: [Sessions]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/IfMatch'}]
      responses: {'202': {$ref: '#/components/responses/Accepted'}, '204': {description: Already closed}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '429': {$ref: '#/components/responses/RateLimited'}, '503': {$ref: '#/components/responses/Unavailable'}, '504': {$ref: '#/components/responses/Timeout'}, default: {$ref: '#/components/responses/Problem'}}
...
  /sessions/{session_id}/events:
    parameters: [{$ref: '#/components/parameters/SessionId'}]
    get:
      operationId: listSessionEvents
      summary: List durable Session events
      tags: [Events]
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses: {'200': {description: Durable event page, content: {application/json: {schema: {$ref: '#/components/schemas/EventPage'}}}}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '410': {$ref: '#/components/responses/Gone'}, '429': {$ref: '#/components/responses/RateLimited'}, '503': {$ref: '#/components/responses/Unavailable'}, default: {$ref: '#/components/responses/Problem'}}
  /sessions/{session_id}/events/stream:
    parameters: [{$ref: '#/components/parameters/SessionId'}]
    get:
      operationId: streamSessionEvents
      summary: Stream durable Session events
      description: Replays durable events strictly after decimal Last-Event-ID (default 0), then polls for new events. Current Session and event disclosure authorization is required throughout. Returns 410 with earliest_sequence/latest_sequence for expired or missing history before streaming; later gaps end the connection with a comment. Heartbeats and retry hints never advance the cursor. Connections have bounded writes, lifetime and per-replica quotas.
      tags: [Events]
      parameters: [{$ref: '#/components/parameters/LastEventId'}]
      responses: {'200': {description: Ordered SSE, content: {text/event-stream: {schema: {type: string}}}}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '410': {$ref: '#/components/responses/Gone'}, '429': {$ref: '#/components/responses/RateLimited'}, '503': {$ref: '#/components/responses/Unavailable'}, default: {$ref: '#/components/responses/Problem'}}
...
components:
  securitySchemes:
    bearerAuth: {type: oauth2, flows: {clientCredentials: {tokenUrl: https://identity.example/oauth2/token, scopes: {ar.admin.read: Read reconciliation status}}}}
  parameters:
    SessionId: {name: session_id, in: path, required: true, schema: {$ref: '#/components/schemas/Id'}}
...
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128, pattern: '^[A-Za-z0-9._~-]+$'}}
    IfMatch: {name: If-Match, in: header, required: true, schema: {type: string, pattern: '^"[1-9][0-9]*"$'}}
    Cursor: {name: cursor, in: query, schema: {type: string, maxLength: 2048}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
    LastEventId: {name: Last-Event-ID, in: header, schema: {type: string, maxLength: 128}}
  responses:
    Accepted: {description: Accepted asynchronous operation, headers: {Location: {schema: {type: string, format: uri-reference}}, Retry-After: {schema: {type: integer, minimum: 1}}}, content: {application/json: {schema: {$ref: '#/components/schemas/Operation'}}}}
    BadRequest: {description: 'Invalid request, cursor, or signal', content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    Unauthorized: {description: Authentication required or credential invalid, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    Forbidden: {description: Permission denied or runtime profile not allowed, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    NotFound: {description: Resource absent or not visible to the caller, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    Conflict: {description: 'State, version, idempotency, or single-writer conflict', content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    Gone: {description: Requested event history is outside retention, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    PayloadTooLarge: {description: Request body exceeds the endpoint limit, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    UnsupportedMediaType: {description: Request media type is unsupported, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    UnprocessableContent: {description: Request is syntactically valid but semantically invalid, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    RateLimited: {description: Caller rate or concurrency budget exhausted, headers: {Retry-After: {schema: {type: integer, minimum: 1}}}, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
...
    Unavailable: {description: Service or required dependency is temporarily unavailable, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    Timeout: {description: A required dependency exceeded its bounded deadline, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    Problem: {description: RFC 7807 problem, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
  schemas:
    Id: {type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'}
    Time: {type: string, format: date-time}
    Session:
      type: object
      additionalProperties: false
      required: [id, state, state_version, execution_generation, agent_runtime_spec_id, runtime_profile_id, workspace_id, created_at, updated_at]
      properties: {id: {$ref: '#/components/schemas/Id'}, state: {enum: [PROVISIONING, READY, ACTIVE, IDLE, SUSPENDED, DEGRADED, CLOSING, CLOSED]}, state_version: {type: integer, minimum: 0}, execution_generation: {type: integer, minimum: 0}, agent_runtime_spec_id: {$ref: '#/components/schemas/Id'}, runtime_profile_id: {$ref: '#/components/schemas/Id'}, workspace_id: {$ref: '#/components/schemas/Id'}, authority_mode: {enum: [local, thinkpixelag]}, authority_issuer: {type: string, maxLength: 128}, current_checkpoint_id: {type: [string, 'null']}, created_at: {$ref: '#/components/schemas/Time'}, updated_at: {$ref: '#/components/schemas/Time'}}
    CreateSession: {type: object, additionalProperties: false, required: [agent_runtime_spec_id, runtime_profile_id, source], properties: {agent_runtime_spec_id: {$ref: '#/components/schemas/Id'}, runtime_profile_id: {$ref: '#/components/schemas/Id'}, source: {type: object, additionalProperties: false, required: [kind], properties: {kind: {enum: [empty, artifact, repository_bundle, thinkpixel_tg]}, reference: {type: string, maxLength: 512}, digest: {type: string, maxLength: 256}}}}}
    SessionPage: {type: object, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/Session'}}, next_cursor: {type: [string, 'null']}}}
...
    EventPage: {type: object, required: [items, earliest_sequence, latest_sequence], properties: {items: {type: array, items: {type: object}}, next_cursor: {type: [string, 'null']}, earliest_sequence: {type: integer, minimum: 1}, latest_sequence: {type: integer, minimum: 0}}}
...
    Operation: {type: object, additionalProperties: false, required: [id, kind, state, created_at], properties: {id: {$ref: '#/components/schemas/Id'}, kind: {type: string}, state: {enum: [PENDING, RUNNING, SUCCEEDED, FAILED]}, resource_location: {type: string, format: uri-reference}, created_at: {$ref: '#/components/schemas/Time'}, completed_at: {type: [string, 'null'], format: date-time}}}
...
    Problem: {type: object, required: [type, title, status], properties: {type: {type: string, format: uri-reference}, title: {type: string, maxLength: 256}, status: {type: integer, minimum: 400, maximum: 599}, detail: {type: string, maxLength: 2048}, instance: {type: string, format: uri-reference}, code: {type: string, maxLength: 80}, request_id: {$ref: '#/components/schemas/Id'}}}